💇‍♀️ Allow parts in project frontmatter
