+++ .changeset/silly-rivers-obey.md
@@ -1,0 +1,5 @@
+---
+'myst-frontmatter': patch
+---
+
+Allow parts in project frontmatter
