+++ packages/myst-frontmatter/src/page/types.ts
@@ -3,16 +3,6 @@ import type { KernelSpec } from '../kernelspec/types.js';
 import type { ProjectAndPageFrontmatter } from '../project/types.js';
 import { PROJECT_AND_PAGE_FRONTMATTER_KEYS } from '../project/types.js';
 
-export const PAGE_KNOWN_PARTS = [
-  'abstract',
-  'summary',
-  'keypoints',
-  'dedication',
-  'epigraph',
-  'data_availability',
-  'acknowledgments',
-];
-
 export const PAGE_FRONTMATTER_KEYS = [
   ...PROJECT_AND_PAGE_FRONTMATTER_KEYS,
   // These keys only exist on the page
@@ -20,9 +10,7 @@ export const PAGE_FRONTMATTER_KEYS = [
   'kernelspec',
   'jupytext',
   'tags',
-  'parts',
   'content_includes_title',
-  ...PAGE_KNOWN_PARTS,
 ];
 
 export type PageFrontmatter = ProjectAndPageFrontmatter & {
@@ -30,7 +18,6 @@ export type PageFrontmatter = ProjectAndPageFrontmatter & {
   kernelspec?: KernelSpec;
   jupytext?: Jupytext;
   tags?: string[];
-  parts?: Record<string, string[]>;
   /** Flag if frontmatter title is duplicated in content
    *
    * Set during initial file/frontmatter load
